Add a script for building U-boot images from Git
By default, it fetches the current state of the mainline U-boot "master"
branch and builds for Firefly ROC-RK3576-PC with opensource TF-A and the
most recent DDR init blob from Rockchip.

USE_BL31=vendor enables the use of the most recent BL31 blob from Rockchip
instead of the opensource TF-A.

Source Git trees and branches can be customized by setting the environment
variables UBOOT_GIT, UBOOT_BRANCH and similarly for TFA and RKBIN.

BOARD=xxx enable builds for the named board, provided that there is
support for it in the specified Git tree. As of July 2025, a wider
selection of boards is available in Jonas Karlman's tree, including
sige5, omni3576, nanopi-m5 and rock-4d.

To avoid redownloading the sources with every build, use the KEEP_SRC
environment variable, taking the following values:
  - yes: Do not touch the downloaded source tree at all, only build it
  - update: Do a "git pull" in the downloaded source to fetch the most
    recent changes before rebuilding
  - no (default): Delete the downloaded sources and do a fresh git clone

Signed-off-by: Alexey Charkov <alchark@gmail.com>