Add script to build a BSP based Linux kernel
Note that Rockchip's BSP sources only include support for their EVB boards
and not for any of the other SBCs.

This script downloads latest device tree files for a number of alternative
boards we want to support (ArmSoM Sige5, Radxa Rock 4D, and FriendlyElec
NanoPi M5). Vendor device tree files for Luckfox Omni3576 are bundled
in this repo, as there is no separate download for them available from the
vendor (only as part of a huge multi-GB SDK archive on Google Drive).

Signed-off-by: Alexey Charkov <alchark@gmail.com>